`2.9.0` Changes
1. **[New Features](#290-new-features)**
2. **[Improvements](#290-improvements)**
3. **[Fixes](#290-fixes)**
4. **[NEW HARDWARE](#290-new-hardware)**

`2.9.0` New Features

+ Added support for USB by implementing an already tested and community driven project - [tinyUSB](https://github.com/hathach/tinyusb)
  + Lightweight and efficient, designed to be small in size and highly efficient in terms of memory and CPU usage
  + All GCC MCUs are supported
    + `STM32`
    + `NXP`
    + `TIVA`
    + `GD32`
  + Following classes added
    + `audio`
    + `bth`
    + `cdc`
    + `dfu`
    + `hid`
    + `midi`
    + `msc`
    + `net`
    + `usbtmc`
    + `vendor`
    + `video`
+ Added support for ETHERNET by implementing an open-source TCP/IP protocol stack library designed for embedded systems and IoT - [CycloneTCP](https://github.com/Oryx-Embedded/CycloneTCP/tree/master)
  + Lightweight and resource-efficient in terms of memory and CPU usage
  + All GCC MCUs are supported
    + `STM32`
    + `NXP`
    + `TIVA`
    + `GD32`
  + Following modules added
    + `coap`
    + `dhcp`
    + `dhcpv6`
    + `dns`
    + `dns_sd`
    + `echo`
    + `ftp`
    + `http`
    + `icecast`
    + `igmp`
    + `ipv4`
    + `ipv6`
    + `lldp`
    + `llmnr`
    + `mdns`
    + `mibs`
    + `modbus`
    + `mqtt`
    + `mqtt_sn`
    + `netbios`
    + `ppp`
    + `smtp`
    + `snmp`
    + `sntp`
    + `syslog`
    + `tftp`
    + `web_socket`
+ Interrupt control added for all architectures and MCUs
+ Added HW initialization sequences for MikroE hardware
  + Sequences added for `USB` and `ETHERNET`
  + Files added as inline interface libraries in `./middleware/hw`
    + **8** files in total for `ETHERNET`
    + **16** files in total for `USB`
+ Added following test examples
  + **USB**
    + `cdc_msc`
    + `hid_generic_inout`
    + `midi`
  + **ETHERNET**
    + `http_client`
    + `ftp_client`

`2.9.0` Improvements

+ SPI low level implementation updated for all architectures
  + Low level `write` and `read` APIs transfer counter changed from `uint16_t` to `size_t`
    + Results in bigger buffers available for single data transfer

`2.9.0` Fixes

+ Fixed pins for USB_UART defined in `dip_socket.h` for packages DIP20 and DIP28 for EasyPIC v8 board
  + `USB_UART_TX` changed from `PC4` to `PC6`
  + `USB_UART_RX` changed from `PC5` to `PC7`
+ Fixed inadequate structure access in following libraries
  + `tft7`
  + `ili9341`
  + `tp_mikroe`

`2.9.0` NEW HARDWARE

+ [SiBRAIN for ATmega2560](https://www.mikroe.com/sibrain-for-atmega2560)
+ [SiBRAIN for ATmega1280](https://www.mikroe.com/sibrain-for-atmega1280)
+ [Clicker 4 for STM32F4](https://www.mikroe.com/clicker-4-for-stm32f4)

+++ targets/riscv/mikroe/gigadevice/src/spi_master/implementation_1/hal_ll_spi_master.c
@@ -484,7 +484,7 @@ uint8_t hal_ll_spi_master_transfer_bare_metal(hal_ll_spi_master_base_handle_t *h
 }
 
 static void hal_ll_spi_master_write_bare_metal( hal_ll_spi_master_base_handle_t *hal_ll_hw_reg, uint8_t *write_data_buffer, size_t write_data_length ) {
-    uint16_t transfer_counter = 0;
+    size_t transfer_counter = 0;
 
     // Write the first data to be transmitted into the SPI_DR register.
     for(transfer_counter = 0; transfer_counter < write_data_length; transfer_counter++) {
@@ -494,7 +494,7 @@ static void hal_ll_spi_master_write_bare_metal( hal_ll_spi_master_base_handle_t 
 }
 
 static void hal_ll_spi_master_read_bare_metal(hal_ll_spi_master_base_handle_t *hal_ll_hw_reg, uint8_t *read_data_buffer, size_t read_data_length, uint8_t dummy_data) {
-    uint16_t transfer_counter = 0;
+    size_t transfer_counter = 0;
 
     // Read the first data to be transmitted into the SPI_DR register.
     for(transfer_counter = 0; transfer_counter < read_data_length; transfer_counter++) {
